* Missed one explict cast...
(bzr r3367.6.3)

+++ launcher/SwitcherModel.cpp
@@ -221,7 +221,7 @@ void SwitcherModel::PrevDetail()
   if (!detail_selection())
     return;
 
-  if (detail_selection_index >= 1)
+  if (detail_selection_index >= (unsigned int) 1)
     detail_selection_index = detail_selection_index - 1;
   else
     detail_selection_index = DetailXids().size() - 1;
